proyecto_final/requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
sentence-transformers==5.3.0
faiss-cpu==1.13.2
pydantic==2.12.5
numpy==2.4.4
